all: extract platform builder from provision
