+++ builder/builder.go
@@ -34,6 +34,23 @@ type Builder interface {
 
 var builders = make(map[string]Builder)
 
+// PlatformBuilder is a builder where administrators can manage
+// platforms (automatically adding, removing and updating platforms).
+type PlatformBuilder interface {
+	PlatformAdd(PlatformOptions) error
+	PlatformUpdate(PlatformOptions) error
+	PlatformRemove(name string) error
+}
+
+// PlatformOptions is the set of options provided to PlatformAdd and
+// PlatformUpdate, in PlatformBuilder.
+type PlatformOptions struct {
+	Name   string
+	Args   map[string]string
+	Input  io.Reader
+	Output io.Writer
+}
+
 // Register registers a new builder in the Builder registry.
 func Register(name string, builder Builder) {
 	builders[name] = builder
