+++ provision/docker/provisioner_test.go
@@ -24,7 +24,6 @@ import (
 	"github.com/fsouza/go-dockerclient/testing"
 	"github.com/tsuru/config"
 	"github.com/tsuru/docker-cluster/cluster"
-	"github.com/tsuru/docker-cluster/storage"
 	"github.com/tsuru/tsuru/app"
 	"github.com/tsuru/tsuru/app/image"
 	"github.com/tsuru/tsuru/builder"
@@ -2125,165 +2124,6 @@ func (s *S) TestProvisionerSleepProcess(c *check.C) {
 	dockerContainer, err = dcli.InspectContainer(cont2.ID)
 	c.Assert(err, check.IsNil)
 	c.Assert(dockerContainer.State.Running, check.Equals, true)
-}
-
-func (s *S) TestProvisionerPlatformAdd(c *check.C) {
-	var requests []*http.Request
-	server, err := testing.NewServer("127.0.0.1:0", nil, func(r *http.Request) {
-		requests = append(requests, r)
-	})
-	c.Assert(err, check.IsNil)
-	defer server.Stop()
-	config.Set("docker:registry", "localhost:3030")
-	defer config.Unset("docker:registry")
-	var p dockerProvisioner
-	err = p.Initialize()
-	c.Assert(err, check.IsNil)
-	p.cluster, _ = cluster.New(nil, &cluster.MapStorage{}, "",
-		cluster.Node{Address: server.URL()})
-	args := make(map[string]string)
-	args["dockerfile"] = "http://localhost/Dockerfile"
-	err = p.PlatformAdd(provision.PlatformOptions{
-		Name:   "test",
-		Args:   args,
-		Output: ioutil.Discard,
-	})
-	c.Assert(err, check.IsNil)
-	c.Assert(len(requests) >= 3, check.Equals, true)
-	requests = requests[len(requests)-3:]
-	c.Assert(requests[0].URL.Path, check.Equals, "/build")
-	queryString := requests[0].URL.Query()
-	c.Assert(queryString.Get("t"), check.Equals, image.PlatformImageName("test"))
-	c.Assert(queryString.Get("remote"), check.Equals, "http://localhost/Dockerfile")
-	c.Assert(requests[1].URL.Path, check.Equals, "/images/localhost:3030/tsuru/test:latest/json")
-	c.Assert(requests[2].URL.Path, check.Equals, "/images/localhost:3030/tsuru/test/push")
-}
-
-func (s *S) TestProvisionerPlatformAddData(c *check.C) {
-	var requests []*http.Request
-	server, err := testing.NewServer("127.0.0.1:0", nil, func(r *http.Request) {
-		requests = append(requests, r)
-	})
-	c.Assert(err, check.IsNil)
-	defer server.Stop()
-	config.Set("docker:registry", "localhost:3030")
-	defer config.Unset("docker:registry")
-	var p dockerProvisioner
-	err = p.Initialize()
-	c.Assert(err, check.IsNil)
-	p.cluster, _ = cluster.New(nil, &cluster.MapStorage{}, "", cluster.Node{Address: server.URL()})
-	dockerfile := "FROM tsuru/java"
-	err = p.PlatformAdd(provision.PlatformOptions{
-		Name:   "test",
-		Args:   map[string]string{"dockerfile": "http://localhost"},
-		Output: ioutil.Discard,
-		Input:  strings.NewReader(dockerfile),
-	})
-	c.Assert(err, check.IsNil)
-	c.Assert(len(requests) >= 3, check.Equals, true)
-	requests = requests[len(requests)-3:]
-	c.Assert(requests[0].URL.Path, check.Equals, "/build")
-	queryString := requests[0].URL.Query()
-	c.Assert(queryString.Get("t"), check.Equals, image.PlatformImageName("test"))
-	c.Assert(queryString.Get("remote"), check.Equals, "")
-	c.Assert(requests[1].URL.Path, check.Equals, "/images/localhost:3030/tsuru/test:latest/json")
-	c.Assert(requests[2].URL.Path, check.Equals, "/images/localhost:3030/tsuru/test/push")
-}
-
-func (s *S) TestProvisionerPlatformAddWithoutArgs(c *check.C) {
-	err := s.p.PlatformAdd(provision.PlatformOptions{Name: "test"})
-	c.Assert(err, check.NotNil)
-	c.Assert(err.Error(), check.Equals, "Dockerfile is required")
-}
-
-func (s *S) TestProvisionerPlatformAddShouldValidateArgs(c *check.C) {
-	args := make(map[string]string)
-	args["dockerfile"] = "not_a_url"
-	err := s.p.PlatformAdd(provision.PlatformOptions{
-		Name:   "test",
-		Args:   args,
-		Output: ioutil.Discard,
-	})
-	c.Assert(err, check.NotNil)
-	c.Assert(err.Error(), check.Equals, "dockerfile parameter must be a URL")
-}
-
-func (s *S) TestProvisionerPlatformAddWithoutNode(c *check.C) {
-	var requests []*http.Request
-	server, err := testing.NewServer("127.0.0.1:0", nil, func(r *http.Request) {
-		requests = append(requests, r)
-	})
-	c.Assert(err, check.IsNil)
-	defer server.Stop()
-	config.Set("docker:registry", "localhost:3030")
-	defer config.Unset("docker:registry")
-	var p dockerProvisioner
-	err = p.Initialize()
-	c.Assert(err, check.IsNil)
-	p.cluster, _ = cluster.New(nil, &cluster.MapStorage{}, "")
-	args := make(map[string]string)
-	args["dockerfile"] = "http://localhost/Dockerfile"
-	err = p.PlatformAdd(provision.PlatformOptions{
-		Name:   "test",
-		Args:   args,
-		Output: ioutil.Discard,
-	})
-	c.Assert(err, check.NotNil)
-}
-
-func (s *S) TestProvisionerPlatformRemove(c *check.C) {
-	registryServer := httptest.NewServer(nil)
-	defer registryServer.Close()
-	u, _ := url.Parse(registryServer.URL)
-	config.Set("docker:registry", u.Host)
-	defer config.Unset("docker:registry")
-	var requests []*http.Request
-	server, err := testing.NewServer("127.0.0.1:0", nil, func(r *http.Request) {
-		requests = append(requests, r)
-	})
-	c.Assert(err, check.IsNil)
-	defer server.Stop()
-	var p dockerProvisioner
-	err = p.Initialize()
-	c.Assert(err, check.IsNil)
-	p.cluster, _ = cluster.New(nil, &cluster.MapStorage{}, "",
-		cluster.Node{Address: server.URL()})
-	var buf bytes.Buffer
-	err = p.PlatformAdd(provision.PlatformOptions{
-		Name:   "test",
-		Args:   map[string]string{"dockerfile": "http://localhost/Dockerfile"},
-		Output: &buf,
-	})
-	c.Assert(err, check.IsNil)
-	err = p.PlatformRemove("test")
-	c.Assert(err, check.IsNil)
-	c.Assert(len(requests) >= 4, check.Equals, true)
-	requests = requests[len(requests)-4:]
-	c.Assert(requests[3].Method, check.Equals, "DELETE")
-	c.Assert(requests[3].URL.Path, check.Matches, "/images/[^/]+")
-}
-
-func (s *S) TestProvisionerPlatformRemoveReturnsStorageError(c *check.C) {
-	registryServer := httptest.NewServer(nil)
-	defer registryServer.Close()
-	u, _ := url.Parse(registryServer.URL)
-	config.Set("docker:registry", u.Host)
-	defer config.Unset("docker:registry")
-	var requests []*http.Request
-	server, err := testing.NewServer("127.0.0.1:0", nil, func(r *http.Request) {
-		requests = append(requests, r)
-	})
-	c.Assert(err, check.IsNil)
-	defer server.Stop()
-	var strg cluster.MapStorage
-	var p dockerProvisioner
-	err = p.Initialize()
-	c.Assert(err, check.IsNil)
-	p.cluster, _ = cluster.New(nil, &strg, "",
-		cluster.Node{Address: server.URL()})
-	err = p.PlatformRemove("test")
-	c.Assert(err, check.NotNil)
-	c.Assert(err, check.DeepEquals, storage.ErrNoSuchImage)
 }
 
 func (s *S) TestProvisionerUnits(c *check.C) {
