+++ provision/provisiontest/fake_provisioner_test.go
@@ -1044,66 +1044,6 @@ func (s *S) TestExecuteCommandOnce(c *check.C) {
 	c.Assert(buf.String(), check.Equals, string(output))
 }
 
-func (s *S) TestExtensiblePlatformAdd(c *check.C) {
-	p := ExtensibleFakeProvisioner{FakeProvisioner: NewFakeProvisioner()}
-	args := map[string]string{"dockerfile": "mydockerfile.txt"}
-	err := p.PlatformAdd(provision.PlatformOptions{Name: "python", Args: args})
-	c.Assert(err, check.IsNil)
-	platform := p.GetPlatform("python")
-	c.Assert(platform.Name, check.Equals, "python")
-	c.Assert(platform.Version, check.Equals, 1)
-	c.Assert(platform.Args, check.DeepEquals, args)
-}
-
-func (s *S) TestExtensiblePlatformAddTwice(c *check.C) {
-	p := ExtensibleFakeProvisioner{FakeProvisioner: NewFakeProvisioner()}
-	args := map[string]string{"dockerfile": "mydockerfile.txt"}
-	err := p.PlatformAdd(provision.PlatformOptions{Name: "python", Args: args})
-	c.Assert(err, check.IsNil)
-	err = p.PlatformAdd(provision.PlatformOptions{Name: "python", Args: nil})
-	c.Assert(err, check.NotNil)
-	c.Assert(err.Error(), check.Equals, "duplicate platform")
-}
-
-func (s *S) TestExtensiblePlatformUpdate(c *check.C) {
-	p := ExtensibleFakeProvisioner{FakeProvisioner: NewFakeProvisioner()}
-	args := map[string]string{"dockerfile": "mydockerfile.txt"}
-	err := p.PlatformAdd(provision.PlatformOptions{Name: "python", Args: args})
-	c.Assert(err, check.IsNil)
-	args["something"] = "wat"
-	err = p.PlatformUpdate(provision.PlatformOptions{Name: "python", Args: args})
-	c.Assert(err, check.IsNil)
-	platform := p.GetPlatform("python")
-	c.Assert(platform.Name, check.Equals, "python")
-	c.Assert(platform.Version, check.Equals, 2)
-	c.Assert(platform.Args, check.DeepEquals, args)
-}
-
-func (s *S) TestExtensiblePlatformUpdateNotFound(c *check.C) {
-	p := ExtensibleFakeProvisioner{FakeProvisioner: NewFakeProvisioner()}
-	err := p.PlatformUpdate(provision.PlatformOptions{Name: "python", Args: nil})
-	c.Assert(err, check.NotNil)
-	c.Assert(err.Error(), check.Equals, "platform not found")
-}
-
-func (s *S) TestExtensiblePlatformRemove(c *check.C) {
-	p := ExtensibleFakeProvisioner{FakeProvisioner: NewFakeProvisioner()}
-	args := map[string]string{"dockerfile": "mydockerfile.txt"}
-	err := p.PlatformAdd(provision.PlatformOptions{Name: "python", Args: args})
-	c.Assert(err, check.IsNil)
-	err = p.PlatformRemove("python")
-	c.Assert(err, check.IsNil)
-	platform := p.GetPlatform("python")
-	c.Assert(platform, check.IsNil)
-}
-
-func (s *S) TestExtensiblePlatformRemoveNotFound(c *check.C) {
-	p := ExtensibleFakeProvisioner{FakeProvisioner: NewFakeProvisioner()}
-	err := p.PlatformRemove("python")
-	c.Assert(err, check.NotNil)
-	c.Assert(err.Error(), check.Equals, "platform not found")
-}
-
 func (s *S) TestFakeProvisionerAddUnit(c *check.C) {
 	app := NewFakeApp("red-sector", "rush", 1)
 	p := NewFakeProvisioner()
